Fix incompatible 'storage_controller' default, conflicting with previously-patched Stemcells
The new default with VirtualBox CPI v0.2.2+ is 'sata', which causes issues for
Bionic Stemcells that have been imported by previous versions of the CPI, that
were using a default if 'ide'.

Indeed, Stemcells OVF definitions are patched by the VirtualBox CPI when being
imported, so that their (Stemcell-based) main disk uses a different storage
controller than the default SCSI.

With version 0.2.0 and prior, the CPI only supported IDE and SCSI, but not
SATA. IDE was behaving better than SCSI for some reason, so the CPI patched
the OVF to use IDE. Bionic Stemcells OVF and prior (Xenial and Trusty) used to
provide an SCSI as InstanceID 3 and an IDE as InstanceID 4. Based on these
assumptions, the CPI patched the OVF blindly from InstanceID 3 (assumed to be
SCSI) to InstanceID 4 (assumed to be IDE).

Then in 2021, the VirtualBox CPI version 0.2.2 has introduced support for SATA
controllers, along with VirtualBox 6+ support. Bionic Stemcells could also use
SATA because the VirtualBox engine supports SATA for a very long time, so the
CPI only needed to patch the Bionic OVF even more, replacing the IDE
controller altogether by a SATA controller. OVF is mere XML after all.

But in 2022, Jammy Stemcells dropped support for IDE controllers. And the
Jammy OVF provides SATA and SCSI controllers, but in reversed order. SATA as
InstanceID 3 and the default SCSI as InstanceID 4. So VirtualBox CPI v0.3.0+
makes more heuristics and assumptions, before patches blindly from InstanceID
4 to 3.

In the end, Importing Jammy Stemcells with VirtualBox CPI 0.2.0 (or prior), or
importing these with a recent CPI using 'ide' as 'storage_controller', is
producing VMs with no bootable disk. (This is hard to spot because the Bosh
Director only experiences a timeout, due to a non-reachable Bosh Agent.)

And existing Bionic Stemcells imported by previous CPI 0.2.0 (or prior) cannot
use SATA. They must use IDE. Same when these have been imported by a more
recent CPI using 'ide' as 'storage_controller'.

Then comes the time to clone the imported Stemcells, and create actual VMs.

When creating VMs, the CPI needs a 'storage_controller' configuration that is
consistent with what was used when importing their Stemcell. But the CPI is
not able to remember what was the configuration used at the time the Stemcell
was imported.

In order to continue using those already-imported Bionic Stemcells, it is
critical to set the 'storage_controller' configuration to 'ide'. But as a
consquence, this setting will also prevent Jammy Stemcells from being properly
imported.

Here with this patch, we needed previously-imported Bionic Stemcells to
work. Plus, the Jammy Stemcells OVF had been messed up when being imported
with the wrong assumptions, so we needed those to be eliminated by the CPI
(which does this only after a successful `bosh create-env' deploy).

So we've had to temporarily switch to 'ide', until we later re-imported and
used the Jammy stemcells with proper 'sata' controller again.